Clean up some compiler warnings.

+++ cmdtab.c
@@ -874,6 +874,6 @@ void *panel_command (void *cmdline)
     }
 #endif /* defined( OPTION_CMDTGT ) */
 
-    return (void*)rc;
+    return (void*)(long)rc;
 }
 /* end panel_command / panel_command_r */
